Add notification for virtual but final methods
If a class implements an interface method, the .net runtime marks this
method as virtual but final, which means it's not overridable. If such
a method should be part of a class (not interface) proxy, we have a
problem: since it's marked as final, it's not "accepted" for our proxy.
BUT, since it's marked as virtual, the configured proxy generation hook
doesn't get any notification.