Only release non-null COM interface pointers
When given a COM object as the target, `CreateInterfaceProxyWith-
TargetInterface` checks whether that target implements the specified
`interfaceToProxy`. It does this by calling `QueryInterface` on the
target, seeing whether it got back a non-null interface pointer, and
releasing the retrieved interface pointer.

However, the retrieved interface pointer must only be released if it
is not a null reference. This commit adds the necessary check for
that.