Explicit interface implementation name collisions
When a proxy is made to implement the same generic interface more than
twice, and that generic interface contains a method that does not use
any generic type parameters, this can be result in a name collision.
This would happen e.g. with a proxy implementing `IObserver<T>` three
times.

DynamicProxy does check for member name collisions, and when such a
one is detected, it switches to explicit interface implementation and
prefixes the member name with the generic interface type name. For the
above example, this will add the following methods to the proxy:

 * OnCompleted
 * IObserver`1.OnCompleted
 * IObserver`1.OnCompleted

It is clearly not sufficient to exclude the concrete type arguments
and use only the number of type parameters. This commit changes the
naming of explicitly implemented interface methods such that they are
unique by including the generic type arguments. For example:

 * OnCompleted
 * IObserver`1[Boolean].OnCompleted
 * IObserver`1[Int32].OnCompleted