Emit invocation types in same module as their associated proxy type
Prior to this commit, DynamicProxy, when asked to proxy a type from a weak-named assembly, would generate the proxy type in the weak-named dynamic assembly, and the associated invocation types in the strong-named dynamic assembly. This results in PEVerify rightfully claiming that both of them are invalid.

Since the assembly of main test project (Castle.Core.Tests) gets strong-named, it's difficult to write tests regarding this issue. Therefore we add a second test project that stays weak-named.