Replace __MonoCS__ for tests affected by mono/mono#8503
On Mono, `DynamicModule.FullyQualifiedName` does not return a fully
qualified name, i.e. one including a file path (as it would on the
CLR, which is the behavior DynamicProxy's `ModuleScope` expects).

There are several tests verifying that dynamic assemblies get stored
to disk correctly. This includes file path and name checks. These can
fail due to the above difference between Mono and the CLR.