Copy custom attributes on ctor params
The only custom attribute that DynamicProxy currently reproduces
on a proxy class' parameters is [ParamArray] (if present on the last
parameter of the base class' ctor).

This commit changes this so DynamicProxy replicates *all* non-
inheritable custom attributes on ctor parameters.

Regarding "non-inheritable": Because proxy ctors have a different
signature than the proxied class' ctors, parameters can never inherit
custom attributes from the base ctor. This means that inheritability
of custom attributes needs to be ignored in the case of proxy ctors.

Also some minor cleanups in `AttributeUtil` while we're looking at it.