Reproduce default parameter values
Reproduction of default parameter values has caused a lot of trouble
in the past and was eventually disabled completely, because
`ParameterBuilder.SetConstant` was deemed too buggy. At that time, it
wasn't clearly understood what bugs exactly caused the problems.

I've spent some time discovering and documenting the relevant bugs
that are present in the CLR, CoreCLR, and on Mono. This commits adds
back the ability to reproduce default parameter values, but this time
with more precise error detection.

This means that on recent runtimes, default parameter values will
stand a good chance of being copied much more faithfully.