Add IInterceptorSelector.SelectInterceptors tests
These verify that the `type` argument passed to `SelectInterceptors`
is equal to the type of the proxy's target. (There are also tests for
`IInvocation.TargetType` in order to ensure consistency between
`IInvocation` and `IInterceptorSelector`.)

One of these tests will fail: When a proxy is created using `Create-
ClassProxy`, `SelectInterceptors` will receive a `type` representing
`System.RuntimeType` instead of the target's type.