Concurrent accesses to `NamingScope.GetUniqueName`
`NamingScope.GetUniqueName` maintains and updates counters in an un-
synchronized `Dictionary<,>`. This writing method is currently called
from a location that allows several concurrent readers.

This changes `BaseProxyGenerator.ObtainProxyType` so that it goes from
a read lock straight to a write lock. There seems little point in
taking an upgradeable lock first if we're going to need a write lock
anyway for all of the work.