Deprecate `Lock` and `ModuleScope`'s public type cache API (#391)
This deprecates `Lock` and directly related types. This is achieved by
replacing internal uses either with `ReaderWriterLockSlim` (upon which
it is based), or with a new but simpler abstraction, `Synchronized-
Dictionary<,>`.

This also deprecates `ModuleScope`'s public API for accessing the
proxy type cache. This is necessary because that API consists of a
`Lock` (which we're deprecating) along with a few thread-unsafe
methods.