Fix tests by actively setting current culture
The tests won't be flaky if we actively control the active culture,
instead of leaving it up to the OS / environment.

We can rely on the `[StringValues]` dictionary behavior formatting
`DateTime` values using a `TypeConverter` fetched via `TypeDescriptor.
GetConverter`, which will fetch `DateTimeConverter`, which will use
`yyyy-MM-dd` format for the invariant culture.

The last bit is documented here:
https://docs.microsoft.com/en-us/dotnet/api/system.componentmodel.datetimeconverter#remarks