Simplify `hook` propagation in `MembersCollector`
Many of `MembersCollector`'s methods have a `hook` parameter so that it
can be forwarded to the three `Add` methods. This wouldn't be necessary
if it were available e.g. via a field.

Simply assigning `hook` to a field in `CollectMembersToProxy` is not
ideal because it worsens the temporal dependency: intermediate methods
now require that `CollectMembersToProxy` be called beforehand.

Assigning `hook` to a field in the constructor would be better, but all
sub-class ctors and their callers would have to be refactored, too.

Turning the `Add` methods (plus intermediates) into local functions of
`CollectMembersToProxy` solves this problem neatly: they gain access to
the `hook` parameter, and the temporal problem evaporates since those
dependent methods can now be called only inside `CollectMembersToProxy`.