Switch to explicit impl for redeclared events
The failing test from the previous commit causes the following metadata
to be emitted:

    .property instance class [mscorlib]System.Action`1<bool> Property()
    {
        .get instance class [mscorlib]System.Action`1<bool>
                         Castle.Proxies.IDerivedProxy::get_Property()
    }

    .property instance class [mscorlib]System.Action Property()
    {
        .get instance class [mscorlib]System.Action
                         Castle.Proxies.IDerivedProxy::get_Property()
    }

Obviously we need to switch to explicit implementation for one of these
events in order to avoid a name collision. We do that by changing the
"collision detection" such that it does not take into account events'
return types; only their name matters.